+++ integrations/crypto-prices/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.0.0-beta.4](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/crypto-prices@1.0.0-beta.3...@kyvejs/crypto-prices@1.0.0-beta.4) (2023-05-19)
+# [1.0.0-beta.5](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/crypto-prices@1.0.0-beta.4...@kyvejs/crypto-prices@1.0.0-beta.5) (2023-06-02)
 
 **Note:** Version bump only for package @kyvejs/crypto-prices
 
+# [1.0.0-beta.4](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/crypto-prices@1.0.0-beta.3...@kyvejs/crypto-prices@1.0.0-beta.4) (2023-05-19)
 
-
-
+**Note:** Version bump only for package @kyvejs/crypto-prices
 
 # [1.0.0-beta.3](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/crypto-prices@1.0.0-beta.2...@kyvejs/crypto-prices@1.0.0-beta.3) (2023-04-24)